+++ app/controllers/all/completed/projects_controller.rb
@@ -1,0 +1,36 @@
+class All::Completed::ProjectsController < ApplicationController
+  after_action :verify_authorized
+  rescue_from ActiveRecord::RecordNotFound, with: :not_found_error
+
+  def index
+    authorize Project, :index?
+    @pager, @projects = pagy(Project.completed)
+
+    pre_fetch_establishments(@projects)
+    pre_fetch_incoming_trusts(@projects)
+  end
+
+  def voluntary
+    authorize Project, :index?
+    @pager, @projects = pagy(Project.completed.voluntary)
+
+    pre_fetch_establishments(@projects)
+    pre_fetch_incoming_trusts(@projects)
+  end
+
+  def sponsored
+    authorize Project, :index?
+    @pager, @projects = pagy(Project.completed.sponsored)
+
+    pre_fetch_establishments(@projects)
+    pre_fetch_incoming_trusts(@projects)
+  end
+
+  private def pre_fetch_establishments(projects)
+    EstablishmentsFetcher.new.call(projects)
+  end
+
+  private def pre_fetch_incoming_trusts(projects)
+    IncomingTrustsFetcher.new.call(projects)
+  end
+end