+++ spec/models/conversion/project_spec.rb
@@ -5,18 +5,18 @@ RSpec.describe Conversion::Project do
     before { mock_successful_api_responses(urn: any_args, ukprn: any_args) }
 
     context "when the project is a voluntary conversion" do
-      let(:project) { create(:voluntary_conversion_project) }
+      let(:project) { create(:conversion_project, sponsor_trust_required: false) }
 
       it "returns the correct route" do
         expect(project.route).to eq :voluntary
       end
     end
 
-    context "when the project is a involuntary conversion" do
-      let(:project) { create(:involuntary_conversion_project) }
+    context "when the project is a sponsored conversion" do
+      let(:project) { create(:conversion_project, sponsor_trust_required: true) }
 
       it "returns the correct route" do
-        expect(project.route).to eq :involuntary
+        expect(project.route).to eq :sponsored
       end
     end
   end