+++ CHANGELOG.md
@@ -12,6 +12,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - users can no longer create a new project with a school URN that already has a
   project in progress, this prevents duplicate projects being created by mistake
 - Removed the "Check the baseline" task from task lists
+- Conversion projects now have a route of voluntary or sponsored
+
+### Added
+
+- a new view that show all in-progress conversion projects that are sponsored at
+  `/projects/all/in-progress/sponsored`
+- a new view that show all in-progress conversion projects that are voluntary at
+  `/projects/all/in-progress/voluntary`
+- a new view that show all completed conversion projects that are sponsored at
+  `/projects/all/in-progress/sponsored`
+- a new view that show all completed conversion projects that are voluntary at
+  `/projects/all/in-progress/voluntary`
 
 ## [Release 17][release-17]